Extract forms into a package with shared widget styling
The four view modules each carried their own byte-identical copy of
FIELD_CLASSES and CHECKBOX_CLASSES, then hand-applied them to 68
individual widgets. A styling change meant finding every copy, and a
field that forgot its attrs rendered unstyled with nothing to catch it.

StyledFormMixin now applies the class by widget type in __init__, using
setdefault so a form that genuinely needs different styling (the roomier
auth screens) keeps its own. Forms declare `widgets` only for the
non-class attributes that are actually specific to them — step, min/max,
rows, placeholder, type=date.

Also adds finance/chart_sections.py so the chart section list lives on
its own instead of Preferences importing it from views_dashboards, which
had forms depending on views for no reason beyond where the constant sat.

Verified by rendering all 13 forms before and after and diffing with
attribute order normalized: exactly one semantic difference, and it is
the intended recognise→recognize from the spelling pass.

  views_settings.py  993 → 568
  views_alerts.py    212 → 121
  views_budgets.py   154 → 106
  views_imports.py   347 → 307
