+++ finance/views_imports.py
@@ -6,28 +6,21 @@ enough to be dangerous when it isn't, so nothing is written until a person has
 looked at the proposed columns and the resulting preview.
 """
 
-from django import forms
 from django.contrib import messages
 from django.shortcuts import redirect
 from django.views.generic import DetailView, FormView, ListView, TemplateView
 
 from .access import FinanceAccessMixin
+from .forms import UploadForm
 from .models import (
-    Account,
     AmountConvention,
     ImportBatch,
     ImportMapping,
     ImportStatus,
-    Institution,
     RecordType,
     RowStatus,
 )
 from .services.importer import ImportError_, commit_batch, parse_batch, stage_upload
-
-FIELD_CLASSES = (
-    "w-full rounded-button border border-border bg-background px-3 py-2 text-sm "
-    "text-text-primary focus:outline-none focus:ring-2 focus:ring-primary"
-)
 
 # Target fields per record type, and whether the import can proceed without them.
 REQUIRED_FIELDS = {
@@ -94,39 +87,6 @@ RECORD_TYPE_GRAIN = {
         "map whichever ones this employer actually itemizes on the stub."
     ),
 }
-
-
-class UploadForm(forms.Form):
-    institution = forms.ModelChoiceField(
-        queryset=Institution.objects.filter(is_active=True),
-        widget=forms.Select(attrs={"class": FIELD_CLASSES}),
-    )
-    account = forms.ModelChoiceField(
-        queryset=Account.objects.filter(is_active=True),
-        required=False,
-        help_text="Required for transactions and balances.",
-        widget=forms.Select(attrs={"class": FIELD_CLASSES}),
-    )
-    record_type = forms.ChoiceField(
-        choices=RecordType.choices,
-        widget=forms.Select(attrs={"class": FIELD_CLASSES}),
-    )
-    csv_file = forms.FileField(
-        widget=forms.ClearableFileInput(attrs={"class": FIELD_CLASSES, "accept": ".csv,text/csv"})
-    )
-
-    def clean(self):
-        cleaned = super().clean()
-
-        if cleaned.get("record_type") in {RecordType.TRANSACTIONS, RecordType.BALANCES} and not cleaned.get("account"):
-            raise forms.ValidationError(
-                "Pick the account these rows belong to — transactions and "
-                "balances cannot be filed without one."
-            )
-
-        return cleaned
-
-
 class ImportListView(FinanceAccessMixin, ListView):
     template_name = "finance/imports/list.html"
     context_object_name = "batches"
