+++ finance/views_settings.py
@@ -5,21 +5,31 @@ affect the ledger, so both people see the same thing.
 
 Preferences are personal — which widgets, which filters. They affect only how
 one person sees that ledger, so they never touch shared data.
+
+The forms these views render live in `finance/forms/`, not here.
 """
 
-from django import forms
 from django.contrib import messages
 from django.db import transaction as db_transaction
 from django.db.models import Count
 from django.shortcuts import get_object_or_404, redirect
 from django.urls import reverse_lazy
 from django.utils import timezone
-from django.utils.text import slugify
 from django.views.generic import CreateView, FormView, ListView, TemplateView, UpdateView
 
 from .access import FinanceAccessMixin
 from .categories_seed import UNCATEGORIZED_SLUG
 from .dates import household_today
+from .forms import (
+    AccountForm,
+    BalanceUpdateForm,
+    CategoryForm,
+    ConnectionForm,
+    InstitutionForm,
+    ManualAccountForm,
+    PreferencesForm,
+    RuleForm,
+)
 from .models import (
     Account,
     AccountConnection,
@@ -41,18 +51,7 @@ from .providers.base import ProviderError
 from .providers.simplefin import claim_access_url
 from .services.categorize import categorize_transactions
 from .services.sync import record_balance_snapshot, sync_connection
-from .services.widgets import WIDGET_CHOICES
 from .views import FinanceView
-from .views_dashboards import CHART_SECTION_CHOICES
-
-FIELD_CLASSES = (
-    "w-full rounded-button border border-border bg-background px-3 py-2 text-sm "
-    "text-text-primary focus:outline-none focus:ring-2 focus:ring-primary"
-)
-
-CHECKBOX_CLASSES = (
-    "h-4 w-4 rounded border-border bg-background text-primary focus:ring-primary"
-)
 
 
 class SettingsHomeView(FinanceView):
@@ -95,41 +94,6 @@ class SettingsHomeView(FinanceView):
             )
 
         return redirect("finance:settings")
-
-
-class InstitutionForm(forms.ModelForm):
-    """Every institution goes through here, whether it ends up connected via
-    SimpleFIN or only ever fed by hand-uploaded statements — CSV import and
-    manual accounts both need an Institution to attach to, and until now the
-    only way to create one was the side effect of connecting a SimpleFIN
-    integration."""
-
-    class Meta:
-        model = Institution
-        fields = ["name", "provider", "owner", "website", "notes", "is_active"]
-        widgets = {
-            "name": forms.TextInput(attrs={"class": FIELD_CLASSES, "placeholder": "Northwestern Mutual"}),
-            "provider": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "owner": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "website": forms.URLInput(attrs={"class": FIELD_CLASSES}),
-            "notes": forms.Textarea(attrs={"class": FIELD_CLASSES, "rows": 3}),
-            "is_active": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-        }
-        help_texts = {
-            "provider": "How data from here normally arrives — SimpleFIN connects "
-            "automatically; CSV/manual means you'll keep it current by hand.",
-        }
-
-    def save(self, commit=True):
-        institution = super().save(commit=False)
-
-        if not institution.slug:
-            institution.slug = slugify(institution.name)[:140]
-
-        if commit:
-            institution.save()
-
-        return institution
 
 
 class InstitutionListView(FinanceAccessMixin, ListView):
@@ -176,41 +140,6 @@ class InstitutionUpdateView(FinanceAccessMixin, UpdateView):
         response = super().form_valid(form)
         messages.success(self.request, f"Updated {self.object.name}.")
         return response
-
-
-class ConnectionForm(forms.Form):
-    """Step one of adding an integration: authenticate.
-
-    No institution field — a single SimpleFIN setup token can cover more than
-    one real institution (that's how SimpleFIN Bridge itself works), so which
-    institution each discovered account belongs to is resolved automatically
-    during sync from the provider's own data, not chosen up front here.
-    """
-
-    label = forms.CharField(
-        label="Name this connection",
-        max_length=120,
-        required=False,
-        widget=forms.TextInput(
-            attrs={"class": FIELD_CLASSES, "placeholder": "Byline + Capital One (joint)"}
-        ),
-        help_text="However you want to recognize this in Settings — it can cover more than one institution.",
-    )
-    setup_token = forms.CharField(
-        label="SimpleFIN setup token",
-        widget=forms.Textarea(
-            attrs={
-                "class": FIELD_CLASSES,
-                "rows": 4,
-                "placeholder": "Paste the token from bridge.simplefin.org",
-                # Never let a browser or password manager retain a bearer
-                # credential from this box.
-                "autocomplete": "off",
-                "spellcheck": "false",
-            }
-        ),
-        help_text="Single use — SimpleFIN will refuse a token that has already been claimed.",
-    )
 
 
 class ConnectionCreateView(FinanceAccessMixin, FormView):
@@ -323,98 +252,6 @@ class ConnectionDetailView(FinanceAccessMixin, TemplateView):
             )
 
         return redirect("finance:connection_detail", pk=connection.pk)
-
-
-class BalanceUpdateForm(forms.Form):
-    """A one-off balance reading, typed in by hand — the lightweight
-    alternative to a full CSV balances import for a single account.
-
-    Deliberately a plain Form, not a ModelForm: saving it does more than set
-    two fields on the Account (see AccountUpdateView._update_balance), it also
-    records an AccountBalanceSnapshot so the reading feeds the same balance
-    history charts a sync or CSV import would.
-    """
-
-    as_of = forms.DateField(
-        label="Balance as of",
-        widget=forms.DateInput(attrs={"class": FIELD_CLASSES, "type": "date"}),
-    )
-    current_balance = forms.DecimalField(
-        label="Balance",
-        max_digits=12,
-        decimal_places=2,
-        widget=forms.NumberInput(attrs={"class": FIELD_CLASSES, "step": "0.01"}),
-        help_text="Signed the same way it reads everywhere else in the app — negative for a debt.",
-    )
-    available_balance = forms.DecimalField(
-        label="Available balance",
-        max_digits=12,
-        decimal_places=2,
-        required=False,
-        widget=forms.NumberInput(attrs={"class": FIELD_CLASSES, "step": "0.01"}),
-    )
-
-
-class AccountForm(forms.ModelForm):
-    class Meta:
-        model = Account
-        fields = [
-            "name", "account_type", "owner", "mask", "debt_reported_positive",
-            "is_active", "include_in_net_worth", "include_in_spending",
-            "sort_order", "notes",
-        ]
-        widgets = {
-            "name": forms.TextInput(attrs={"class": FIELD_CLASSES}),
-            "account_type": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "owner": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "mask": forms.TextInput(attrs={"class": FIELD_CLASSES}),
-            "sort_order": forms.NumberInput(attrs={"class": FIELD_CLASSES}),
-            "notes": forms.Textarea(attrs={"class": FIELD_CLASSES, "rows": 2}),
-            "debt_reported_positive": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-            "is_active": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-            "include_in_net_worth": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-            "include_in_spending": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-        }
-
-
-class ManualAccountForm(forms.ModelForm):
-    """For an account with no connection — the mortgage, the 401(k), anything
-    only ever kept current by a CSV import or by hand. A connected account's
-    institution is set by the sync that discovered it, so this form (and its
-    institution field) is create-only; editing an existing account goes
-    through AccountForm instead."""
-
-    class Meta:
-        model = Account
-        fields = [
-            "institution", "name", "account_type", "owner", "mask",
-            "current_balance", "balance_as_of",
-            "debt_reported_positive", "include_in_net_worth", "include_in_spending",
-            "notes",
-        ]
-        widgets = {
-            "institution": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "name": forms.TextInput(attrs={"class": FIELD_CLASSES, "placeholder": "401(k)"}),
-            "account_type": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "owner": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "mask": forms.TextInput(attrs={"class": FIELD_CLASSES}),
-            "current_balance": forms.NumberInput(attrs={"class": FIELD_CLASSES, "step": "0.01"}),
-            "balance_as_of": forms.DateTimeInput(attrs={"class": FIELD_CLASSES, "type": "datetime-local"}),
-            "notes": forms.Textarea(attrs={"class": FIELD_CLASSES, "rows": 2}),
-            "debt_reported_positive": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-            "include_in_net_worth": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-            "include_in_spending": forms.CheckboxInput(attrs={"class": CHECKBOX_CLASSES}),
-        }
-        help_texts = {
-            "current_balance": "Optional — leave blank and set it later with a CSV "
-            "balance import if you don't have a figure handy right now.",
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["institution"].queryset = Institution.objects.filter(is_active=True)
-        self.fields["current_balance"].required = False
-        self.fields["balance_as_of"].required = False
 
 
 class AccountCreateView(FinanceAccessMixin, CreateView):
@@ -537,51 +374,6 @@ class RuleListView(FinanceAccessMixin, ListView):
         return redirect("finance:rules")
 
 
-class RuleForm(forms.ModelForm):
-    """A CategoryRule, editable from Settings rather than only ever created
-    sight-unseen via the review queue's "always" checkbox (which only ever
-    writes a CONTAINS match on the merchant name). The model already
-    supports contains/starts-with/exact/regex matching, optionally scoped
-    to one account or an amount range — this just exposes it."""
-
-    class Meta:
-        model = CategoryRule
-        fields = [
-            "pattern", "match_type", "category", "account",
-            "min_amount", "max_amount", "priority", "notes",
-        ]
-        widgets = {
-            "pattern": forms.TextInput(attrs={"class": FIELD_CLASSES, "placeholder": "netflix"}),
-            "match_type": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "category": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "account": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "min_amount": forms.NumberInput(attrs={"class": FIELD_CLASSES, "step": "0.01"}),
-            "max_amount": forms.NumberInput(attrs={"class": FIELD_CLASSES, "step": "0.01"}),
-            "priority": forms.NumberInput(attrs={"class": FIELD_CLASSES}),
-            "notes": forms.TextInput(attrs={"class": FIELD_CLASSES}),
-        }
-        help_texts = {
-            "pattern": "Matched against the transaction's raw description, case-insensitively.",
-            "match_type": "“Contains” is the most forgiving choice when an exact match is unlikely.",
-            "account": "Leave unset to apply everywhere.",
-            "min_amount": "Signed, inclusive lower bound. Leave blank for no lower bound.",
-            "max_amount": "Signed, inclusive upper bound. Leave blank for no upper bound.",
-            "priority": "Lower runs first. The first matching rule wins.",
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["category"].queryset = Category.objects.filter(
-            is_active=True, children__isnull=True
-        ).select_related("parent").alphabetical()
-        self.fields["account"].queryset = Account.objects.filter(is_active=True)
-        self.fields["account"].required = False
-        self.fields["account"].empty_label = "Every account"
-        self.fields["min_amount"].required = False
-        self.fields["max_amount"].required = False
-        self.fields["notes"].required = False
-
-
 class RuleCreateView(FinanceAccessMixin, CreateView):
     template_name = "finance/settings/rule_form.html"
     form_class = RuleForm
@@ -600,77 +392,6 @@ class RuleCreateView(FinanceAccessMixin, CreateView):
             f"“{rule.pattern}” now goes to {rule.category}.",
         )
         return super().form_valid(form)
-
-
-class CategoryForm(forms.ModelForm):
-    """Create or rename a category. The slug is set once, at creation, and
-    never changes after — several system categories (Uncategorized, Internal
-    Transfer, Credit Card Payment) are looked up by slug in code, and a
-    handful of other slugs are what CategoryRule/MerchantCategoryMemo
-    matching keys off of, so renaming a category must not disturb it."""
-
-    class Meta:
-        model = Category
-        fields = ["name", "parent", "kind", "sort_order", "description"]
-        widgets = {
-            "name": forms.TextInput(attrs={"class": FIELD_CLASSES, "placeholder": "Groceries"}),
-            "parent": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "kind": forms.Select(attrs={"class": FIELD_CLASSES}),
-            "sort_order": forms.NumberInput(attrs={"class": FIELD_CLASSES}),
-            "description": forms.TextInput(attrs={"class": FIELD_CLASSES}),
-        }
-        help_texts = {
-            "parent": "Leave unset for a top-level category. Nesting goes at most three levels deep.",
-            "kind": "A subcategory must share its parent's kind.",
-            "sort_order": "Lower sorts first among siblings.",
-            "description": "Steers the classifier — say what belongs here and what doesn't.",
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.fields["parent"].queryset = Category.objects.filter(is_active=True)
-        self.fields["parent"].required = False
-        self.fields["parent"].empty_label = "No parent (top-level)"
-        self.fields["description"].required = False
-
-    def clean_parent(self):
-        parent = self.cleaned_data.get("parent")
-
-        if parent and self.instance.pk:
-            # Without this guard, picking one of a category's own descendants
-            # as its new parent creates a cycle — full_path/depth walk parent
-            # links and would loop forever.
-            node = parent
-            while node is not None:
-                if node.pk == self.instance.pk:
-                    raise forms.ValidationError(
-                        "A category can't be nested under one of its own subcategories."
-                    )
-                node = node.parent
-
-        return parent
-
-    def save(self, commit=True):
-        category = super().save(commit=False)
-
-        if not category.slug:
-            category.slug = self._unique_slug(category.name)
-
-        if commit:
-            category.save()
-
-        return category
-
-    def _unique_slug(self, name):
-        base = slugify(name)[:90] or "category"
-        slug = base
-        suffix = 1
-
-        while Category.objects.filter(slug=slug).exclude(pk=self.instance.pk).exists():
-            suffix += 1
-            slug = f"{base}-{suffix}"
-
-        return slug
 
 
 class CategoryListView(FinanceAccessMixin, ListView):
@@ -826,152 +547,6 @@ class CategoryDeleteView(FinanceAccessMixin, TemplateView):
             f"Deleted {name}. {moved} transaction{'s' if moved != 1 else ''} moved to {reassign_to.full_path}.",
         )
         return redirect("finance:categories")
-
-
-class PreferencesForm(forms.ModelForm):
-    widgets_selected = forms.MultipleChoiceField(
-        label="Homepage widgets",
-        choices=WIDGET_CHOICES,
-        required=False,
-        widget=forms.CheckboxSelectMultiple(attrs={"class": CHECKBOX_CLASSES}),
-    )
-    # Populated by the drag-and-drop reorder script (finance/js/widget-reorder.js)
-    # as a comma-separated slug list reflecting the on-screen row order —
-    # every WIDGET_CHOICES slug, not just the checked ones, so unchecking and
-    # rechecking a widget doesn't lose its place in the list.
-    widget_order = forms.CharField(required=False, widget=forms.HiddenInput())
-    chart_sections_selected = forms.MultipleChoiceField(
-        label="Charts tab sections",
-        choices=CHART_SECTION_CHOICES,
-        required=False,
-        widget=forms.CheckboxSelectMultiple(attrs={"class": CHECKBOX_CLASSES}),
-    )
-    # Same reorder pattern as widget_order, one section down.
-    chart_section_order_input = forms.CharField(required=False, widget=forms.HiddenInput())
-    accounts_selected = forms.ModelMultipleChoiceField(
-        label="Accounts to show",
-        queryset=Account.objects.filter(is_active=True),
-        required=False,
-        widget=forms.CheckboxSelectMultiple(attrs={"class": CHECKBOX_CLASSES}),
-        help_text="Leave all unchecked to show every account.",
-    )
-    budgets_selected = forms.ModelMultipleChoiceField(
-        label="Budgets to show",
-        queryset=Budget.objects.filter(is_active=True),
-        required=False,
-        widget=forms.CheckboxSelectMultiple(attrs={"class": CHECKBOX_CLASSES}),
-        help_text="Leave all unchecked to show every active budget.",
-    )
-
-    class Meta:
-        model = UserPreference
-        fields = ["recent_transaction_count"]
-        widgets = {
-            "recent_transaction_count": forms.NumberInput(
-                attrs={"class": FIELD_CLASSES, "min": 1, "max": 50}
-            )
-        }
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-        all_slugs = [slug for slug, _ in WIDGET_CHOICES]
-        all_section_slugs = [slug for slug, _ in CHART_SECTION_CHOICES]
-
-        if self.instance.pk:
-            self.fields["widgets_selected"].initial = self.instance.widgets
-            self.fields["accounts_selected"].initial = self.instance.homepage_account_ids
-            self.fields["budgets_selected"].initial = self.instance.homepage_budget_ids
-            # The chosen widgets first, in the order the person last arranged
-            # them, then anything not yet chosen — so a widget shipped after
-            # they last saved still shows up, at the end rather than nowhere.
-            ordered = list(self.instance.widgets) + [
-                slug for slug in all_slugs if slug not in self.instance.widgets
-            ]
-            # Filtered to slugs that still exist: a section retired from
-            # CHART_SECTION_CHOICES must not linger in an already-saved
-            # preference forever, or this KeyErrors below on the label
-            # lookup the moment someone opens Preferences.
-            saved_sections = [
-                slug for slug in self.instance.chart_section_order
-                if slug in all_section_slugs
-            ]
-            self.fields["chart_sections_selected"].initial = saved_sections
-            ordered_sections = saved_sections + [
-                slug for slug in all_section_slugs if slug not in saved_sections
-            ]
-        else:
-            ordered = all_slugs
-            ordered_sections = all_section_slugs
-
-        self.fields["widget_order"].initial = ",".join(ordered)
-        self._ordered_slugs = ordered
-
-        self.fields["chart_section_order_input"].initial = ",".join(ordered_sections)
-        self._ordered_section_slugs = ordered_sections
-
-    def ordered_widget_rows(self):
-        """(slug, label, checked) in the order the reorder UI should render them."""
-        checked = set(self.fields["widgets_selected"].initial or [])
-        labels = dict(WIDGET_CHOICES)
-
-        return [(slug, labels[slug], slug in checked) for slug in self._ordered_slugs]
-
-    def ordered_chart_section_rows(self):
-        """(slug, label, checked) for the Charts tab section reorder UI."""
-        checked = set(self.fields["chart_sections_selected"].initial or [])
-        labels = dict(CHART_SECTION_CHOICES)
-
-        return [
-            (slug, labels[slug], slug in checked) for slug in self._ordered_section_slugs
-        ]
-
-    def save(self, commit=True):
-        preference = super().save(commit=False)
-
-        # The hidden field carries the drag-and-drop order for every known
-        # widget; filtering it down to what's checked (rather than reading
-        # WIDGET_CHOICES' fixed order) is what lets a person actually control
-        # placement instead of always getting it reset to the declared order.
-        chosen = set(self.cleaned_data["widgets_selected"])
-        known_slugs = {slug for slug, _ in WIDGET_CHOICES}
-        ordered_slugs = [
-            slug
-            for slug in self.cleaned_data["widget_order"].split(",")
-            if slug in known_slugs
-        ]
-        # A missing or corrupted order (JS disabled, stale form) falls back
-        # to the declared order rather than dropping every widget.
-        if not ordered_slugs:
-            ordered_slugs = [slug for slug, _ in WIDGET_CHOICES]
-
-        preference.homepage_widgets = [slug for slug in ordered_slugs if slug in chosen]
-
-        chosen_sections = set(self.cleaned_data["chart_sections_selected"])
-        known_section_slugs = {slug for slug, _ in CHART_SECTION_CHOICES}
-        ordered_section_slugs = [
-            slug
-            for slug in self.cleaned_data["chart_section_order_input"].split(",")
-            if slug in known_section_slugs
-        ]
-        if not ordered_section_slugs:
-            ordered_section_slugs = [slug for slug, _ in CHART_SECTION_CHOICES]
-
-        preference.chart_sections = [
-            slug for slug in ordered_section_slugs if slug in chosen_sections
-        ]
-
-        preference.homepage_account_ids = [
-            account.pk for account in self.cleaned_data["accounts_selected"]
-        ]
-        preference.homepage_budget_ids = [
-            budget.pk for budget in self.cleaned_data["budgets_selected"]
-        ]
-
-        if commit:
-            preference.save()
-
-        return preference
 
 
 class PreferencesView(FinanceAccessMixin, UpdateView):
